+++ semcore/data-table/__tests__/index.test.tsx
@@ -525,6 +525,7 @@ describe('DataTable', () => {
                   return {
                     value: index,
                     active: [1, 2].includes(index),
+                    fontWeight: 'normal',
                     collapse: {
                       children: (
                         <Box
@@ -589,6 +590,7 @@ describe('DataTable', () => {
                   return {
                     value: index,
                     active: [1, 2].includes(index),
+                    fontWeight: 'normal',
                     collapse: {
                       children: (
                         <DataTable data={data} zIndex={2}>