+++ semcore/accordion/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [5.36.1] - 2024-06-26
+
+### Changed
+
+- `Accordion.Item.Toggle` can have only `h*` tag, `h3` tag is default.
+
+### Added
+
+- Flex component `Accordion.Item.ToggleButton` with role `button`.
+
 ## [5.36.0] - 2024-06-26
 
 ### Changed